+++ cppbuild/src/ImStack.hpp
@@ -8,20 +8,20 @@ class ImStack
 
 protected:
   ImFrame*** m_frames;
-  uint8_t m_nwaves;
-  uint8_t m_nz;
+  int m_nwaves;
+  int m_nz;
 
 public:
   ImStack();
-  ImStack(uint8_t nw, uint8_t nz);
+  ImStack(int nw, int nz);
   ~ImStack();
-  void insert(ImFrame** frames, uint8_t startW, uint8_t startZ, uint8_t nw, uint8_t nz, uint8_t order);
-  void insert(ImFrame* frame, uint8_t w, uint8_t z);
+  void insert(ImFrame** frames, int startW, int startZ, int nw, int nz, int order);
+  void insert(ImFrame* frame, int w, int z);
   ImStack* zprojection();
   void divide(int d);
-  ImFrame* frame(uint8_t w, uint8_t z){ return m_frames[w][z]; }
-  uint8_t nwaves(){ return m_nwaves; }
-  uint8_t nz(){ return m_nz; }
+  ImFrame* frame(int w, int z){ return m_frames[w][z]; }
+  int nwaves(){ return m_nwaves; }
+  int nz(){ return m_nz; }
 
 };
